moved get_timestamp_in_seconds method in datetime preset handler into utils function
get_timestamp_in_seconds function used in multiple places so moved to
utils.py. Added tests and fixed bugs for convert_timestamp_to_seconds
function
