+++ lib/openstack_query/handlers/presets/preset_handler_datetime.py
@@ -1,10 +1,11 @@
-from datetime import timedelta, datetime
-from custom_types.openstack_query.aliases import PresetToValidPropsMap
+from datetime import datetime
 
 from enums.query.query_presets import QueryPresets, QueryPresetsDateTime
-from openstack_query.handlers.presets.preset_handler_base import PresetHandlerBase
 
-from exceptions.missing_mandatory_param_error import MissingMandatoryParamError
+from custom_types.openstack_query.aliases import PresetToValidPropsMap
+
+from openstack_query.utils import get_timestamp_in_seconds
+from openstack_query.handlers.presets.preset_handler_base import PresetHandlerBase
 
 
 class PresetHandlerDateTime(PresetHandlerBase):
@@ -29,32 +30,6 @@ class PresetHandlerDateTime(PresetHandlerBase):
             QueryPresetsDateTime.YOUNGER_THAN_OR_EQUAL_TO: self._prop_younger_than_or_equal_to,
         }
 
-    @staticmethod
-    def _get_current_time():
-        return datetime.now()
-
-    def _get_timestamp_in_seconds(
-        self, days: int = 0, hours: int = 0, minutes: int = 0, seconds: float = 0
-    ):
-        """
-        Method which takes a number of days, hours, minutes, and seconds - and calculates the total seconds
-        :param days: (Optional) number of days
-        :param hours: (Optional) number of hours
-        :param minutes: (Optional) number of minutes
-        :param seconds: (Optional) number of seconds
-        """
-        if all(arg == 0 for arg in [days, hours, minutes, seconds]):
-            raise MissingMandatoryParamError(
-                "requires at least 1 argument for function to be non-zero"
-            )
-
-        current_time = self._get_current_time().timestamp()
-        prop_time_in_seconds = timedelta(
-            days=days, hours=hours, minutes=minutes, seconds=seconds
-        ).total_seconds()
-
-        return current_time - prop_time_in_seconds
-
     def _prop_older_than(
         self,
         prop: str,
@@ -74,7 +49,7 @@ class PresetHandlerDateTime(PresetHandlerBase):
         :param prop_timestamp_fmt: (Optional) timestamp format of prop value (default: yyyy-MM-ddTHH:mm:ssZ)
         """
         prop_timestamp = datetime.strptime(prop, prop_timestamp_fmt).timestamp()
-        given_timestamp = self._get_timestamp_in_seconds(days, hours, minutes, seconds)
+        given_timestamp = get_timestamp_in_seconds(days, hours, minutes, seconds)
 
         return prop_timestamp > given_timestamp
 
@@ -119,9 +94,7 @@ class PresetHandlerDateTime(PresetHandlerBase):
         :param prop_timestamp_fmt: (Optional) timestamp format of prop value (default: yyyy-MM-ddTHH:mm:ssZ)
         """
         prop_datetime = datetime.strptime(prop, prop_timestamp_fmt).timestamp()
-        return prop_datetime < self._get_timestamp_in_seconds(
-            days, hours, minutes, seconds
-        )
+        return prop_datetime < get_timestamp_in_seconds(days, hours, minutes, seconds)
 
     def _prop_older_than_or_equal_to(
         self,
