+++ tests/lib/openstack_query/handlers/test_preset_handler_datetime.py
@@ -3,15 +3,12 @@ from datetime import datetime
 from unittest.mock import patch
 from parameterized import parameterized
 
-from nose.tools import raises
-
 from openstack_query.handlers.presets.preset_handler_datetime import (
     PresetHandlerDateTime,
 )
 from tests.lib.openstack_query.mocks.mocked_props import MockProperties
 
 from enums.query.query_presets import QueryPresetsDateTime
-from exceptions.missing_mandatory_param_error import MissingMandatoryParamError
 
 
 class PresetHandlerGenericTests(unittest.TestCase):
@@ -47,34 +44,6 @@ class PresetHandlerGenericTests(unittest.TestCase):
         Tests that handler supports all generic query presets
         """
         self.assertIsNotNone(self.instance._get_mapping(preset, MockProperties.PROP_1))
-
-    @parameterized.expand(
-        [
-            ("1 day = 86400 seconds", 1, 0, 0, 0, 86400),
-            ("12 hours = 43200 seconds", 0, 12, 0, 0, 43200),
-            ("2 days, 30 minutes = 174600 seconds", 2, 0, 30, 0, 174600),
-            ("20 seconds", 0, 0, 0, 20, 20),
-        ]
-    )
-    @patch(
-        "openstack_query.handlers.presets.preset_handler_datetime.PresetHandlerDateTime._get_current_time"
-    )
-    def test_get_timestamp_in_seconds(
-        self, name, days, hours, minutes, seconds, total_seconds, mock_current_datetime
-    ):
-        mock_current_datetime.return_value = datetime(2023, 6, 4, 10, 30, 0)
-        out = self.instance._get_timestamp_in_seconds(days, hours, minutes, seconds)
-
-        mock_current_datetime.assert_called_once()
-
-        expected_timestamp = (
-            mock_current_datetime.return_value.timestamp() - total_seconds
-        )
-        assert out == expected_timestamp
-
-    @raises(MissingMandatoryParamError)
-    def test_get_timestamp_in_seconds_invalid(self):
-        self.instance._get_timestamp_in_seconds(days=0, hours=0, minutes=0, seconds=0)
 
     @parameterized.expand(
         [
@@ -130,9 +99,7 @@ class PresetHandlerGenericTests(unittest.TestCase):
             ),
         ]
     )
-    @patch(
-        "openstack_query.handlers.presets.preset_handler_datetime.PresetHandlerDateTime._get_current_time"
-    )
+    @patch("openstack_query.utils.get_current_time")
     def test_prop_older_than(
         self,
         name,
@@ -212,9 +179,7 @@ class PresetHandlerGenericTests(unittest.TestCase):
             ),
         ]
     )
-    @patch(
-        "openstack_query.handlers.presets.preset_handler_datetime.PresetHandlerDateTime._get_current_time"
-    )
+    @patch("openstack_query.utils.get_current_time")
     def test_prop_older_than_or_equal_to(
         self,
         name,
@@ -294,9 +259,7 @@ class PresetHandlerGenericTests(unittest.TestCase):
             ),
         ]
     )
-    @patch(
-        "openstack_query.handlers.presets.preset_handler_datetime.PresetHandlerDateTime._get_current_time"
-    )
+    @patch("openstack_query.utils.get_current_time")
     def test_prop_younger_than(
         self,
         name,
@@ -376,9 +339,7 @@ class PresetHandlerGenericTests(unittest.TestCase):
             ),
         ]
     )
-    @patch(
-        "openstack_query.handlers.presets.preset_handler_datetime.PresetHandlerDateTime._get_current_time"
-    )
+    @patch("openstack_query.utils.get_current_time")
     def test_prop_younger_than_or_equal_to(
         self,
         name,
